requirements.txt
asgiref==3.2.10
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.0
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
